Refactor: Create 컴포넌트 리팩토링
- 앨범 커버 이미지, 슬라이더 브레이크포인트, 에러 메시지 등을 상수로 분리
- 사용하는 타입 분리
- 사용하는 스타일 분리
- AlbumCover 컴포넌트 분리
- 앨범명, 부제목 작성 & 제출 시 빈 값 검사 과정 훅으로 분리
