+++ src/router/index.tsx
@@ -5,10 +5,9 @@ import { createBrowserRouter } from 'react-router-dom';
 
 import { ROUTES_PATH } from '@Constants/routes';
 
-
 const Login = lazy(() => import('@Pages/Login'));
 const Main = lazy(() => import('@Pages/Main'));
-const Create = lazy(() => import('@Pages/Create'));
+const Create = lazy(() => import('@Pages/Create/Create'));
 const Individual = lazy(() => import('@Pages/IndividualPage/Individual'));
 const Decoration = lazy(() => import('@Pages/Decoration'));
 const QrCode = lazy(() => import('@Pages/QrCode'));
