+++ src/types/create.ts
@@ -1,0 +1,9 @@
+export interface CoverDivProps {
+  $isSelected?: boolean;
+}
+
+export type ErrorSetter = (val: boolean) => void;
+
+export type ValidateInput = (input: string, errorSetter: ErrorSetter, errorMessage: string) => boolean;
+
+export type FocusType = 'title' | 'subTitle';
